requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.8
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.2
httplib2==0.14.0
Markdown==3.4.1
oauthlib==3.1.0
PyJWT==2.5.0
pymacaroons==0.13.0
python-dateutil==2.7.3
requests==2.22.0
simplejson==3.16.0
pathlib==1.0.1
mutagen==1.46.0
shortuuid==1.0.9
drf-spectacular==0.24.2
drf-spectacular-sidecar==2022.11.1
django-database-view==0.3.0
